# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateAnnouncementRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  handlers.CreateCarouselRequest:
    properties:
      imageUrl:
        type: string
      linkUrl:
        type: string
      sortOrder:
        type: integer
      title:
        type: string
    type: object
...
  handlers.UpdateAnnouncementRequest:
    properties:
      content:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
...
  models.Announcement:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      status:
        description: active, inactive
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  models.Carousel:
    properties:
      id:
        type: integer
      imageUrl:
        type: string
      linkUrl:
        type: string
      sortOrder:
        type: integer
      status:
        description: active, inactive
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: PetShop 宠物商店 API 文档 - 提供宠物管理、购物车、订单管理、用户管理等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PetShop API
  version: "1.0"
paths:
  /api/admin/announcement:
    delete:
      consumes:
      - application/json
      description: 根据ID删除公告
      parameters:
      - description: 公告ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功消息
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 公告不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 删除公告
      tags:
      - 公告管理
    put:
      consumes:
      - application/json
      description: 更新现有公告信息
      parameters:
      - description: 公告更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateAnnouncementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的公告
          schema:
            $ref: '#/definitions/models.Announcement'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 公告不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新公告
      tags:
      - 公告管理
  /api/admin/announcements:
    get:
      consumes:
      - application/json
      description: 获取所有公告
      produces:
      - application/json
      responses:
        "200":
          description: 公告列表
          schema:
            items:
              $ref: '#/definitions/models.Announcement'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取公告列表
      tags:
      - 公告管理
    post:
      consumes:
      - application/json
      description: 创建新的公告
      parameters:
      - description: 公告创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAnnouncementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功的公告
          schema:
            $ref: '#/definitions/models.Announcement'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 创建公告
      tags:
      - 公告管理
  /api/admin/carousel:
    delete:
      consumes:
      - application/json
      description: 根据ID删除轮播图
      parameters:
      - description: 轮播图ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功消息
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 轮播图不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 删除轮播图
      tags:
      - 轮播图管理
    put:
      consumes:
      - application/json
      description: 更新现有轮播图信息
      parameters:
      - description: 轮播图信息
        in: body
        name: carousel
        required: true
        schema:
          $ref: '#/definitions/models.Carousel'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的轮播图
          schema:
            $ref: '#/definitions/models.Carousel'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 轮播图不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新轮播图
      tags:
      - 轮播图管理
  /api/admin/carousels:
    get:
      consumes:
      - application/json
      description: 获取所有轮播图
      produces:
      - application/json
      responses:
        "200":
          description: 轮播图列表
          schema:
            items:
              $ref: '#/definitions/models.Carousel'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取轮播图列表
      tags:
      - 轮播图管理
    post:
      consumes:
      - application/json
      description: 创建新的轮播图
      parameters:
      - description: 轮播图创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCarouselRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功的轮播图
          schema:
            $ref: '#/definitions/models.Carousel'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 创建轮播图
      tags:
      - 轮播图管理
  /api/admin/config:
    post:
      consumes:
      - application/json
      description: 设置系统配置值。如果key为"inventory_threshold"，会更新库存预警阈值
      parameters:
      - description: 系统配置设置请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SetSystemConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 设置成功消息
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 设置系统配置
      tags:
      - 系统配置
  /api/admin/configs:
    get:
      consumes:
      - application/json
      description: 获取所有系统配置项
      produces:
      - application/json
      responses:
        "200":
          description: 系统配置列表
          schema:
            items:
              $ref: '#/definitions/models.SystemConfig'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取系统配置列表
      tags:
      - 系统配置
  /api/admin/inventory/adjust:
    post:
      consumes:
      - application/json
      description: 调整指定商品的库存数量（增加或减少）
      parameters:
      - description: 库存调整请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.InventoryAdjustRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的商品信息
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 商品不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 调整商品库存
      tags:
      - 库存管理
  /api/admin/inventory/alerts:
    get:
      consumes:
      - application/json
      description: 获取库存低于预警阈值的商品列表
      produces:
      - application/json
      responses:
        "200":
          description: 库存预警列表
          schema:
            items:
              $ref: '#/definitions/models.InventoryAlert'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取库存预警
      tags:
      - 库存管理
  /api/admin/inventory/logs:
    get:
      consumes:
      - application/json
      description: 获取所有库存变更记录
      produces:
      - application/json
      responses:
        "200":
          description: 库存变更日志列表
          schema:
            items:
              $ref: '#/definitions/models.Inventory'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取库存变更日志
      tags:
      - 库存管理
  /api/admin/order:
    get:
      consumes:
      - application/json
      description: 根据订单ID获取订单详细信息
      parameters:
      - description: 订单ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 订单详情
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 订单不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取订单详情
      tags:
      - 订单管理
    put:
      consumes:
      - application/json
      description: 更新订单的状态
      parameters:
      - description: 状态更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的订单
          schema:
            $ref: '#/definitions/models.Order'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 订单不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新订单状态
      tags:
      - 订单管理
  /api/admin/order/refund:
    post:
      consumes:
      - application/json
      description: 为订单处理退款并恢复库存
      parameters:
      - description: 退款请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 退款成功消息
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 订单不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 处理订单退款
      tags:
      - 订单管理
  /api/admin/orders:
    get:
      consumes:
      - application/json
      description: 获取所有订单列表，支持按状态筛选
      parameters:
      - description: 订单状态筛选 (pending, paid, shipped, delivered, cancelled, refunding,
          refunded)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 订单列表
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取订单列表
      tags:
      - 订单管理
  /api/admin/product:
    delete:
      consumes:
      - application/json
      description: 将商品标记为已删除（软删除）
      parameters:
      - description: 商品ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功消息
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 商品不存在
          schema:
            type: string
        "409":
          description: 商品有待处理订单
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 删除商品
      tags:
      - 商品管理
    get:
      consumes:
      - application/json
      description: 根据ID获取商品详细信息
      parameters:
      - description: 商品ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 商品详情
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 商品不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取商品详情
      tags:
      - 商品管理
    put:
      consumes:
      - application/json
      description: 更新现有商品信息
      parameters:
      - description: 商品更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的商品
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 商品不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 更新商品
      tags:
      - 商品管理
  /api/admin/products:
    get:
      consumes:
      - application/json
      description: 获取所有商品列表（不包括已删除的）
      produces:
      - application/json
      responses:
        "200":
          description: 商品列表
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取商品列表
      tags:
      - 商品管理
    post:
      consumes:
      - application/json
      description: 创建新商品
      parameters:
      - description: 商品创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功的商品
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 创建商品
      tags:
      - 商品管理
  /api/admin/stats/hot-products:
    get:
      consumes:
      - application/json
      description: 获取销量最高的商品列表
      parameters:
      - default: 10
        description: 返回数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 热销商品列表
          schema:
            items:
              $ref: '#/definitions/models.HotProduct'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取热销商品
      tags:
      - 销售统计
  /api/admin/stats/sales:
    get:
      consumes:
      - application/json
      description: 获取销售统计数据，支持日、周、月维度
      parameters:
      - default: day
        description: 统计周期 (day:最近7天, week:最近4周, month:最近6月)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 销售统计数据
          schema:
            items:
              $ref: '#/definitions/models.SalesStat'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取销售统计
      tags:
      - 销售统计
...
  /api/admin/user/reset-password:
    post:
      consumes:
      - application/json
      description: 重置指定用户的密码
      parameters:
      - description: 密码重置请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功消息
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: 用户不存在
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 重置用户密码
      tags:
      - 用户管理
  /api/admin/users:
    get:
      consumes:
      - application/json
      description: 获取所有用户列表
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "401":
          description: 未授权
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取用户列表
      tags:
      - 用户管理
...
securityDefinitions:
  APITokenAuth:
    description: 使用 API Token 进行认证，用于 Open API 访问
    in: header
    name: X-API-Token
    type: apiKey
  BearerAuth:
    description: 使用 JWT Token 进行认证，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"